backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==0.28.1
pillow==9.5.0
requests==2.28.0
gtts==2.3.0
python-dotenv==1.0.0 